packages/ai-ml-service/requirements.txt
fastapi==0.111.*
uvicorn[standard]==0.30.*
pandas==2.2.*
prophet==1.1.5      # heavy; install once
pydantic==2.*
pydantic-settings==2.3.*
joblib==1.4.*
torch==2.3.*
scikit-learn==1.5.*
statsmodels==0.14.*
psycopg2-binary==2.9.*
APScheduler==3.10.*
tensorflow==2.16.*